fixing -Wall errors
git-svn-id: https://qthreads.googlecode.com/svn/trunk@853 9dbd018f-d1da-d342-4c11-58c73dfe904b

+++ test/time_gcd.c
@@ -70,7 +70,7 @@ int main ()
     qtimer_stop(timer2);
     for (i=0;i<BIGNUM;i++) {
 	if (answer1[i] != answer2[i]) {
-	    printf("ERROR! %i\n", i);
+	    printf("ERROR! %lu\n", (unsigned long)i);
 	}
     }
     printf("new secs: %f\n", qtimer_secs(timer1));
